Change ChreApiTestService::copyString parameter to size_t
Bug: 258074235
Test: build the nanoapp

+++ apps/test/common/chre_api_test/inc/chre_api_test_manager.h
@@ -117,7 +117,7 @@ class ChreApiTestService final
    * @param source              the source string
    * @param maxChars            the maximum number of chars
    */
-  void copyString(char *destination, const char *source, uint32_t maxChars);
+  void copyString(char *destination, const char *source, size_t maxChars);
 };
 
 /**
